pyproject.toml
[project]
name = "a8p5-hands-on-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
backend = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "sseclient-py>=1.8.0",
    "uvicorn>=0.35.0",
]
config = [
    "hydra-core>=1.3.2",
]
dev = [
    "ipykernel>=6.30.0",
    "jq>=1.10.0",
]
framework = [
    "haystack-ai>=2.15.2",
    "ollama-haystack>=3.4.0",
    "qdrant-haystack>=9.2.0",
]
frontend = [
    "gradio>=5.38.2",
]
preprocess = [
    "nltk>=3.9.1",
    "pypdf>=5.8.0",
    "python-pptx>=1.0.2",
    "sentence-transformers>=5.0.0",
]
test = [
    "pandas>=2.3.1",
]
